pyproject.toml
[project]
name = "delta-lake-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "delta-spark>=4.0.0",
    "deltalake>=1.0.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "nbformat>=5.10.4",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "polars>=1.31.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
]

[project.scripts]
delta-lake-health = "delta_lake_health:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
